# add memory-mapped file loading and prepare exclusive file locking