--- a/src-tauri/Cargo.toml
+++ b/src-tauri/Cargo.toml
@@ -59,6 +59,7 @@ num_cpus = "1.17.0"
 mimalloc = "0.1.48"
 image_hasher = "3.0.0"
 regex = "1.11.2"
+memmap2 = "0.9.8"
 
 [build-dependencies]
 tauri-build = { version = "2.4", features = [] }