--- a/src-tauri/Cargo.toml
+++ b/src-tauri/Cargo.toml
@@ -60,6 +60,8 @@ mimalloc = "0.1.48"
 image_hasher = "3.0.0"
 regex = "1.11.2"
 memmap2 = "0.9.8"
+fs2 = "0.4.3"
+io = "0.0.2"
 
 [build-dependencies]
 tauri-build = { version = "2.4", features = [] }