This change sets up the groundwork for opening and reading files

- Load files into memory using memmap2
- Introduce future exclusive locking (currently working on this)
- Foundation to prevent files from being modified during thumbnail generation

related: https://github.com/CyberTimon/RapidRAW/issues/307 - https://github.com/CyberTimon/RapidRAW/issues/330