@CyberTimon 2025-09-24
Thanks - I’ll take a look this evening and test it.
By the way, why was the “under 20 MB” text changed to “under 40 MB”? The Windows and macOS installers - which most users download - are still well under 20 MB (about 18 MB):
https://github.com/CyberTimon/RapidRAW/releases/download/v1.3.13/01_RapidRAW_v1.3.13_windows_x64.exe