@riccio8 2025-09-24
> Time for a 1.14.0

I’m testing it right now by opening 2,365 images (about one third are RAW) from my NAS.
The version I modified finished in 36 seconds (still too long), while the current release version is still loading (2 minutes and 43 seconds so far).
Once you go above 1,000 images, the user has to wait around 20 seconds before thumbnails start to appear. I'll improve that