@CyberTimon 2025-09-24
**Performance tests**

- **260 A7iii RAW images**  
  - main: 20s  
  - pr: 18s  

- **170 Mavic 4 Pro RAW images**  
  - main: 14s  
  - pr: 16s  

- **110 Air 3 RAW images**  
  - main: 16s  
  - pr: 16s  

@riccio8 I didn’t notice any significant speed difference. I tested across multiple folders, and performance was basically the same.  
Do you think the speedup is only noticeable with folders containing 1k+ images?

@william-sy Do you notice a speed difference? Thanks

Also, thank you for the feedback regarding the logo. I’ll think about it for 1-2 days and then release the final version alongside **v1.4.0**.