@riccio8 2025-09-24
> **Performance tests**
> 
> - **260 A7iii RAW images**  
>   - main: 20s  
>   - pr: 18s  
> 
> - **170 Mavic 4 Pro RAW images**  
>   - main: 14s  
>   - pr: 16s  
> 
> - **110 Air 3 RAW images**  
>   - main: 16s  
>   - pr: 16s  
> 
> @riccio8 I didn’t notice any significant speed difference. I tested across multiple folders, and performance was basically the same.  
> Do you think the speedup is only noticeable with folders containing 1k+ images?
> 
> @william-sy Do you notice a speed difference? Thanks
> 
> Also, thank you for the feedback regarding the logo. I’ll think about it for 1-2 days and then release the final version alongside **v1.4.0**.
> 

Probably, because I 'm iterating on chunks of 1000 elements using rayon, so before 1000 it's kind of the same thing. And if u already opened the folder he would use the ore existing img, without mapping in memory, and when u open a folder o file, u're blocking exclusively because this reduces the possibility of the file to change while it is open. There would be buffer overflows because a fixed amount of swap equivalent to the file size is allocated.

I tried as I said on a Nas in my local network and it seems to be a lot faster, I'll test again, tomorrow if u want I'll send a screen record. Thx for feedbacks