@riccio8 2025-09-25
<img width="3839" height="2057" alt="immagine" src="https://github.com/user-attachments/assets/6bca9cc2-a9a6-4d82-a078-4c94533286f1" />

@CyberTimon there’s a bug in both versions: the program can’t find and list files and directory structures after like 2 minutes. Probably the issue is in the recursive directory function. I’ll look into it; maybe that’s what happened to some users.