@CyberTimon 2025-09-25
I was doing some in-depth testing on your PR today and ran into a few issues that we'll need to sort out before this can be merged.

Here’s what I found:

1.  **Intermittent Crash on Rapid Navigation:** When I open the library and navigate through images very quickly, the app sometimes crashes with a stack buffer overrun. It's a bit tricky to reproduce, but it seems to happen when closing and opening images fast.
    ```
    2025-09-25 17:27:45 [ERROR] PANIC! in src\\file_management.rs:433
    ...
    (exit code: 0xc0000409, STATUS_STACK_BUFFER_OVERRUN)
    ```

2.  **Panic on Re-opening an Image:** This one is 100% reproducible. If you load an image, close it, and then immediately load the same image again, it panics.
    ```
    2025-09-25 17:30:51 [ERROR] PANIC! in src\\main.rs:1139
    ```

3.  **High Res View Bug:** When I hit 'F' to go full-screen, I can feel the lag as it loads the high-res image, but it never actually displays it - it just stays on the low-res preview. Zooming in also doesn't seem to load the detailed tiles (in normal editor mode).