@CyberTimon 2025-09-25
Thank you for looking into it & yes - definitely. The logging system is great, I really like it. :)

> The problem is that I implemented trylockexclusive. I did this so that no other process OR OTHER THREAD could modify it, in order to prevent buffer overflows and other more serious and risky bugs. What happens is that it does not have an arc<mutex, so when it is opened and closed quickly, there are virtual threads that cooperate, but part of the file is still open in a thread, so trylockexclusive returns that it failed to lock all the resources. As soon as the thread finishes, it returns the missing part, and when you reopen it, it is like opening the complete image.

Makes sense, thanks for explaining

> I had the same problem, but it also crashed with the current release 1.3.13

Okay that's strange. Any error?