@CyberTimon 2025-09-25
> No, probably because it wasn't an error like a crash, but more like a bug in the logic, it's really strange.
Fixed problem n2, it was really easy, when do u plan to make a new release?

> For problem n2: [lock_shared](https://doc.rust-lang.org/std/fs/struct.File.html#method.try_lock_shared)

Thanks for fixing it! I had to work on something else today, so I couldn’t really focus on RapidRAW. Tomorrow is also pretty limited for me, but I think Saturday, the 27th, would be a good day for releasing 1.4.0 👍.

> What do you think about implementing a section that other software doesn't have, for image recovery for example? There are great crates like [image_recovery](https://docs.rs/image-recovery/latest/image_recovery/) or [oidn](https://docs.rs/oidn/latest/oidn/#example) that are very useful, and maybe even something like [rusty_sr](https://lib.rs/gh/millardjn/rusty_sr/rusty_sr). Would that work?

I’ve already looked into these kinds of features, and I definitely plan to implement something along those lines soon. Most likely I’ll write the algorithms myself rather than rely on wrappers, since those often create cross platform issues.

Thanks again! I’ll retest your PR soon, possibly tomorrow evening.