@riccio8 2025-09-25
> Sorry, but even at first glance this issue looks like linking OIDN across all three major platforms won’t be straightforward: https://github.com/Twinklebear/oidn-rs/issues/28
> 
> I might take another look at denoising this weekend to see if I can find a native option. In the meantime, if you’d like to experiment, feel free to create a branch, try things out, and report back.

I didn't see that. I'll try something in the next few days probably tomorrow evening otherwise not before 27, I'll tell you. And for the graphics part? Do u want me to start a discussion ok this arguments or I keep writing here?