@william-sy 2025-09-26
Note sure what to think of things this is the latest release I get booted up by the looks of things (mac aarch):
https://github.com/CyberTimon/RapidRAW/actions/runs/17923539645

Not sure what happens here: 
https://github.com/CyberTimon/RapidRAW/actions/runs/17924135981

But for me after this merge it does not boot up any longer.