@CyberTimon 2025-09-26
Can you try the main branch, manually remove these lines and test again:
<img width="1305" height="498" alt="image" src="https://github.com/user-attachments/assets/43225af5-d24a-4074-9fc7-ef348c936337" />