@CyberTimon 2025-09-26
https://github.com/CyberTimon/RapidRAW/discussions/150

will respond later, but check this out