@CyberTimon 2025-09-26
> Also 1 image was sliced up before it rendered. Did not mange to reproduce just yet.

This is normal when running it from source, as sometimes the react useeffects attach at the wrong time because of the hot reloading module. You can reproduce it probably by pressing F5 many times. It doesn't happen in releases.

Are there any more new issues and could you test @riccio8's PR?